+++ lib/db/src/schema/index.ts
@@ -9,3 +9,4 @@ export * from "./locationProfiles";
 export * from "./contentLibrary";
 export * from "./generatedContent";
 export * from "./lifeEvents";
+export * from "./authTokens";